+++ app/models/imports/geo_names.rb
@@ -1,0 +1,5 @@
+module Imports
+  module GeoNames
+    require Rails.root.join("app/lib/imports/geonames/region_record_connector")
+  end
+end