+++ spec/factories/regions.rb
@@ -3,11 +3,12 @@ FactoryBot.define do
     sequence(:public_id) { Nanoid.generate(size: Region::PUBLIC_ID_LENGTH) }
     sequence(:name) { |n| "Region #{n}" }
     sequence(:slug) { |n| "region-#{n}" }
-    region_type { Region::REGION_TYPES[:country] }
+    region_kind { Region::REGION_KINDS[:country] }
+    country_code { "ID" }
     status { Region::STATUSES[:active] }
     parent { nil }
     summary { nil }
     description { nil }
-    external_ref { nil }
+    center { nil }
   end
 end